@@ -1,111 +1,110 @@
 {
-  "name": "adminto-react",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@fullcalendar/bootstrap": "^5.10.1",
-    "@fullcalendar/core": "^5.10.1",
-    "@fullcalendar/daygrid": "^5.10.1",
-    "@fullcalendar/interaction": "^5.10.1",
-    "@fullcalendar/list": "^5.10.1",
-    "@fullcalendar/react": "^5.10.1",
-    "@fullcalendar/timegrid": "^5.10.1",
-    "@hookform/resolvers": "^2.8.8",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/html-to-draftjs": "^1.4.0",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.7.13",
-    "@types/react-albus": "^2.0.8",
-    "@types/react-datepicker": "^4.3.4",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-draft-wysiwyg": "^1.13.4",
-    "@types/react-redux": "^7.1.22",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/react-table": "^7.7.9",
-    "@types/react-text-mask": "^5.4.11",
-    "@types/redux-saga": "^0.10.5",
-    "apexcharts": "^3.33.1",
-    "axios": "^0.25.0",
-    "axios-mock-adapter": "^1.20.0",
-    "bootstrap": "^5.1.3",
-    "chart.js": "^3.7.1",
-    "classnames": "^2.3.1",
-    "deni-react-treeview": "^1.1.11",
-    "draft-js": "^0.11.7",
-    "draftjs-to-html": "^0.9.1",
-    "easymde": "^2.16.1",
-    "env-cmd": "^10.1.0",
-    "eslint-config-react-app": "6",
-    "google-maps-react": "^2.0.6",
-    "html-to-draftjs": "^1.5.0",
-    "i18next": "^21.6.10",
-    "i18next-browser-languagedetector": "^6.1.3",
-    "jsvectormap": "1.3.3",
-    "jwt-decode": "^3.1.2",
-    "nouislider-react": "^3.4.1",
-    "react": "^18.1.0",
-    "react-albus": "^2.0.0",
-    "react-apexcharts": "^1.3.9",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-bootstrap": "^2.1.1",
-    "react-bootstrap-typeahead": "^6.0.0-alpha.8",
-    "react-chartjs-2": "^4.0.1",
-    "react-countdown": "^2.3.2",
-    "react-countup": "^6.1.1",
-    "react-datepicker": "^4.6.0",
-    "react-dom": "^17.0.2",
-    "react-draft-wysiwyg": "^1.14.7",
-    "react-dropzone": "^12.0.4",
-    "react-hook-form": "^7.25.1",
-    "react-i18next": "^11.15.3",
-    "react-image-lightbox": "^5.1.4",
-    "react-joyride": "^2.3.2",
-    "react-nestable": "^2.0.0",
-    "react-redux": "^7.2.6",
-    "react-responsive-masonry": "^2.1.4",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.1",
-    "react-select": "^5.2.2",
-    "react-simplemde-editor": "^5.0.2",
-    "react-sweetalert2": "^0.5.1",
-    "react-table": "^7.7.0",
-    "react-text-mask": "^5.4.3",
-    "redux-saga": "^1.1.3",
-    "sass": "^1.49.0",
-    "simplebar-react": "^2.3.6",
-    "typescript": "^4.4.2",
-    "web-vitals": "^2.1.0",
-    "yarn": "^1.22.18",
-    "yup": "^0.32.11"
-  },
-  "scripts": {
-    "start:dev": "env-cmd -f .env.development react-scripts start",
-    "start:prod": "env-cmd -f .env.production react-scripts start",
-    "build:dev": "env-cmd -f .env.development react-scripts build",
-    "build:prod": "env-cmd -f .env.production react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "format": "prettier --write 'src/**/*.{ts,tsx}'"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/draftjs-to-html": "^0.8.1",
-    "@types/react": "^17.0.45",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "prettier": "2.5.1"
-  }
+    "name": "adminto-react",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "@fullcalendar/bootstrap": "^6.1.7",
+        "@fullcalendar/core": "^6.1.7",
+        "@fullcalendar/daygrid": "^6.1.7",
+        "@fullcalendar/interaction": "^6.1.7",
+        "@fullcalendar/list": "^6.1.7",
+        "@fullcalendar/react": "^6.1.7",
+        "@fullcalendar/timegrid": "^6.1.7",
+        "@hookform/resolvers": "^3.1.0",
+        "@reduxjs/toolkit": "^1.9.5",
+        "@testing-library/jest-dom": "^5.14.1",
+        "@testing-library/react": "^14.0.0",
+        "@testing-library/user-event": "^14.4.3",
+        "@types/html-to-draftjs": "^1.4.0",
+        "@types/jest": "^29.5.1",
+        "@types/node": "^20.1.4",
+        "@types/react-albus": "^2.0.8",
+        "@types/react-datepicker": "^4.11.2",
+        "@types/react-dom": "^18.2.4",
+        "@types/react-draft-wysiwyg": "^1.13.4",
+        "@types/react-redux": "^7.1.22",
+        "@types/react-router-dom": "^5.3.3",
+        "@types/react-table": "^7.7.9",
+        "@types/react-text-mask": "^5.4.11",
+        "@types/redux-saga": "^0.10.5",
+        "apexcharts": "^3.40.0",
+        "axios": "^1.4.0",
+        "axios-mock-adapter": "^1.21.4",
+        "bootstrap": "5.1.3",
+        "chart.js": "^4.3.0",
+        "classnames": "^2.3.1",
+        "deni-react-treeview": "^1.1.11",
+        "draft-js": "^0.11.7",
+        "draftjs-to-html": "^0.9.1",
+        "easymde": "^2.18.0",
+        "env-cmd": "^10.1.0",
+        "eslint-config-react-app": "7.0.1",
+        "google-maps-react": "^2.0.6",
+        "html-to-draftjs": "^1.5.0",
+        "i18next": "^22.4.15",
+        "i18next-browser-languagedetector": "^7.0.1",
+        "jsvectormap": "^1.5.3",
+        "jwt-decode": "^3.1.2",
+        "nouislider-react": "^3.4.1",
+        "react": "^18.2.0",
+        "react-albus": "^2.0.0",
+        "react-apexcharts": "^1.4.0",
+        "react-beautiful-dnd": "^13.1.0",
+        "react-bootstrap": "2.1.1",
+        "react-bootstrap-typeahead": "6.0.0-alpha.8",
+        "react-chartjs-2": "^5.2.0",
+        "react-countdown": "^2.3.2",
+        "react-countup": "^6.4.2",
+        "react-datepicker": "^4.11.0",
+        "react-dom": "^18.2.0",
+        "react-draft-wysiwyg": "^1.15.0",
+        "react-dropzone": "^14.2.3",
+        "react-hook-form": "^7.43.9",
+        "react-i18next": "^12.2.2",
+        "react-image-lightbox": "^5.1.4",
+        "react-joyride": "^2.5.4",
+        "react-nestable": "^2.0.0",
+        "react-redux": "^8.0.5",
+        "react-responsive-masonry": "^2.1.4",
+        "react-router-dom": "^6.11.1",
+        "react-scripts": "^5.0.1",
+        "react-select": "^5.7.3",
+        "react-simplemde-editor": "^5.2.0",
+        "react-sweetalert2": "^0.5.2",
+        "react-table": "^7.8.0",
+        "react-text-mask": "^5.5.0",
+        "sass": "^1.62.1",
+        "simplebar-react": "^3.2.4",
+        "typescript": "^5.0.4",
+        "web-vitals": "^3.3.1",
+        "yarn": "^1.22.18",
+        "yup": "^1.1.1"
+    },
+    "scripts": {
+        "start:dev": "env-cmd -f .env.development react-scripts start",
+        "start:prod": "env-cmd -f .env.production react-scripts start",
+        "build:dev": "env-cmd -f .env.development react-scripts build",
+        "build:prod": "env-cmd -f .env.production react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject",
+        "format": "prettier --write 'src/**/*.{ts,tsx}'"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "devDependencies": {
+        "@types/draftjs-to-html": "^0.8.1",
+        "@types/react": "^18.2.6",
+        "@types/react-beautiful-dnd": "^13.1.2",
+        "prettier": "2.8.8"
+    }
 }
